+++ src/ElementFinder.php
@@ -286,12 +286,15 @@
 
 
     /**
+     * @deprecated
+     * @see element
      * Alias of ElementFinder::query
      *
      * @param string $expression
      * @return \DOMNodeList
      */
     public function node($expression) {
+      trigger_error('Deprecated. Use element() method', E_USER_DEPRECATED);
       return $this->query($expression);
     }